modemfwd: replace DLOG with modemfwd-specific verbose log
This allows us to turn up the log verbosity on official images to
get a handle on what's going on.

The new log can be turned on manually but is always on when the
image is a debug image.

BUG=chromium:980372
TEST=deploy, build trybot image and deploy

Change-Id: I701740b699ca9021ce8311372486b154325a95e6
Reviewed-on: https://chromium-review.googlesource.com/1688137
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>