power: Fix up power supplies for 4.19+ kernels
In older kernels, /sys/class/power_supply/<instance/type had fairly
dynamic contents that indicated whether or not the power supply was
USB dual role. In 4.19 kernels, this file is very static, and for
Chrome OS EC USB ports, just has "USB" in it. Instead, a new "usb_type"
file is present that has the dual-role info in it. It comes in a form
like this, with brackets around the selected one:

Unknown SDP DCP CDP C PD [PD_DRP] BrickID

If the type file has "USB" in it, attempt to access usb_type. If it works
and what's selected is PD_DRP, then use this as an indication of a
dual-role port.

BUG=b:123541187
TEST=Plug in another chromebook via type-C, observe dropdown in
Settings>Power. Tested on Hatch and Kohaku.

Change-Id: Ifaac85c81842f6643be92d681ce69550bf5e0e43
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1684359
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>