kerberos: Add Kerberos configuration validator
The Kerberos daemon should not accept all options in a Kerberos
configuration, but only a whitelisted subset. Thus, this CL adds a class
to validate configuration data. See
https://web.mit.edu/kerberos/krb5-1.12/doc/admin/conf_files/krb5_conf.html
for a description of the format and a list of all options.

It was necessary to implement a parser for the configuration since that
functionality is not publicly exposed by the MIT krb5 library. The
parser is loosely based on prof_parse.c:parse_line(), see
https://github.com/krb5/krb5/blob/master/src/util/profile/prof_parse.c#L277,
but with a less C-like implementation and only checked array access for
security.

BUG=chromium:978806
TEST=Unit tests

Change-Id: I33f75b4f4441dfb5a1fbb81b0b01f72a0a59c71e
Reviewed-on: https://chromium-review.googlesource.com/1687882
Tested-by: Lutz Justen <ljusten@chromium.org>
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>