cros-disks: Allow unpartitioned devices to be formatted to NTFS
It is currently impossible for a user to distinguish between an
unpartitioned device and a device with a single partition in the files
app. Thus, it can seem like formatting to ntfs fails inconsistently on
devices with a single filesystem, as mkfs.ntfs will not allow writing a
filesystem to the raw device by default. Add --force to allow it to.

BUG=chromium:632988
TEST=test_that $DUT e:platform_CrosDisks.*
TEST=manually format an unpartitioned USB drive through the files app

Change-Id: I46e294e6362c9d7617ddb26179664f584ffd29cc
Reviewed-on: https://chromium-review.googlesource.com/1687004
Tested-by: Austin Tankiang <austinct@chromium.org>
Commit-Ready: Austin Tankiang <austinct@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Anand Mistry <amistry@chromium.org>