cryptohome: Add out-of-process mounter.
We're finally there. All the pieces are in place to use an
out-of-process mounter for Guest sessions. This CL implements it, but
does not yet use it.

The implementation lives in a new OutOfProcessMountHelper class, which
will require minimal modifications to the Mount class. This was the
point of introducing the MountHelper class in the first place.

cryptohome communicates with the helper via stdin and stdout.
Once the helper is launched, it will wait to read the system salt from
cryptohome, and at that point perform the ephemeral mount. After that
it will signal cryptohome letting it know that the mount was
successful. cryptohome will write to the helper's stdin ("poke") when
it's time to tear down the ephemeral mount, and the helper will also
tear down the mounts if it's signalled with SIGTERM.

The helper uses a message loop to be able to watch stdin for
cryptohome's poke. This allows both watching for SIGTERM and waiting
for cryptohome's poke with shared functionality.

BUG=chromium:985492
TEST=New unit tests, existing unit tests pass.

Change-Id: I4509857e2ff466c0669edfdd080f9608f45d9ba2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1895909
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>