ml: Get memory usage from proc/[pid]/status
The previous code parses /proc/[pid]/totmaps to obtain the memory usage,
but this is deprecated and has been removed upstream (crbug.com/979781).
In the modification, we instead parse the /proc/[pid]/status and use
VmRSS+VmSwap as the memory usage.

The new method does not use the hand-coded indices so the result
should be more robust.

The new number reported will be a bit different from the previous one
because it will include the shared memory. This should be fine because
it is the relative sizes of the numbers that matter, rather the absolute
values.

BUG=chromium:979781
TEST=Added util_test.cc and ran emerge-pyro ml
TEST=Also ran cros_workon_make --board=$BOARD $PACKAGE_NAME --test
TEST=Confirmed it fixed the crash in crbug.com/979781 on device

Change-Id: I7fbfb6a1daa2db046bab9644d954404272c54330
Reviewed-on: https://chromium-review.googlesource.com/1692241
Tested-by: Honglin Yu <honglinyu@chromium.org>
Commit-Ready: Honglin Yu <honglinyu@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>