login: Add --crash-loop-before if we might be in a crash loop
In session_manager, when launching a user's Chrome process, see if a
quick failure of that process will cause the user to be logged out
(because their previous Chrome sessions have failed rapid, that is,
because we are in a crash loop). If so, add --crash-loop-before to the
Chrome execution. This will cause crash_reporter to upload any crashes
immediately instead of writing them to the user directory, which will be
locked after the next crash.

This is the last functional change needed to enable crash-loop mode!

BUG=chromium:857317
TEST=Repeatedly "kill -SEGV" chrome. Eventually got an UploadSingleCrash
crash report in crash-staging, just as session_manager logged me out.

Change-Id: I18b5d5b150bc7a1e0d56fb815864c6a45bd9cdcb
Reviewed-on: https://chromium-review.googlesource.com/1696906
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Ready: Ian Barkley-Yeung <iby@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>