crosh/debugd: Don't anonymize MACs in network_diag
There's an inconsistency with how we anonymize MACs for feedback reports
in that it is done in network_diag and then again in the anonymizer in
debugd & Chrome. This removes one layer of the redaction so it can be
then done (later) in a singular place so that we have consistency
amongst all redacted MAC addresses in the logs. For now it is still done
in two places (debugd & Chrome).

This also changes the default to NOT anonymize the output because on
manual invocation, that is likely more desirable and we control the
automated invocations of it which have been updated with the new
parameter.

BUG=chromium:673939
TEST=Manually verified redaction in feedback report

Change-Id: Icf8f61586fe249d066442bf14f30cd207b07af05
Reviewed-on: https://chromium-review.googlesource.com/1697526
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>