dlcservice: Modify install API to take protobuf list of DLC ids
The previous API for install took a single DLC id for the install. Now
the API for install will take a list of DLC ids.

Changes to the |InstallResult| and |DlcModuleInfo| are for
|DlcModuleList| to be leveraged by the |OnStatusUpdateSignal()| as well.
These changes will not break backwards compatability as code is not in
production use. No users would be affected.

A future work to do here is to handle scenarios where not all DLC
modules may successfully be installed.

BUG=chromium:905075
TEST=FEATURES="test" emerge-$BOARD dlcservice dlcservice-client system_api
TEST=emerge-$BOARD dlcservice dlcservice-client system_api
TEST=cros deploy $TEST_IP dlcservice dlcservice-client system_api
TEST=tast run $TEST_IP platform.DLCService

Change-Id: I3053312d31680b7965436dc8dc42c82a09a95974
Reviewed-on: https://chromium-review.googlesource.com/1696527
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Ready: Jae Hoon Kim <kimjae@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Xiaochu Liu <xiaochu@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>