crash: Write out flag file to indicate crash init is done
There's a corresponding test for this which has been flakey for a long
time and a recent change was made to determine better the root cause
and it's because we are unable to find the boot line in the logs;
likely because they have been flooded with other problems and the
message no longer exists. The purpose of the check was to ensure
crash_reporter actually ran its init and set the core_pattern. Just
checking core_pattern isn't reliable enough since it gets tweaked in
other tests. This adds a new flag file written out by crash_reporter
which we then check the timestamp in the test to verify it was
written after boot.

BUG=chromium:940608
TEST=test_that logging_CrashSender passes (fails if flag is missing/old)

Change-Id: I7bdbc6c1431f7af487823fe5d0edd372338261d7
Reviewed-on: https://chromium-review.googlesource.com/1699434
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tim Zheng <timzheng@chromium.org>