shill: wifi: Use supplicant disconnect reason code for metrics
Prior to this change, the metrics for disconnect reason codes did not
come from supplicant, but instead were parsed from netlink
messages. Not only was this an unnecessary duplication of
functionality, but it meant that WiFi code and the corresponding
disconnect metric were not necessarily in sync.

This change modifies the situation such that this metric is updated by
disconnect reasons received from supplicant. The primary functional
difference this change makes is to stop converting
kReasonCodeReserved* disconnect reasons into kReasonCodeInvalid, as
the previous behavior made it more difficult to understand the actual
distribution of disconnect reasons, without providing much benefit
(also, kReasonCodeInvalid is the maximum value of a uint16_t, and so
is outside the range of the reason code enum to begin with).

BUG=None
TEST=-All unit tests are passing.
     -Manual testing with extra logging added in
      Metrics::Notify80211Disconnect.

Change-Id: Ie57ea63e31b60ee0c73bfb63594f1ad99c2d65ce
Reviewed-on: https://chromium-review.googlesource.com/1704343
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>