shill: metrics: Rename WiFiStatusType to WiFiReasonType
The WiFiStatusType enum is used exclusively for representing the type
of wifi disconnect reason. Since the standard uses "reason" to mean a
disconnect reason and "status" to mean an auth/assoc status code, it
makes sense to be consistent and use "reason" in this case. This
change also fixes a bug in which kStatusCodeMax was used instead of
kReasonCodeMax for the maximum value of a disconnect reason
code (which likely occurred in the first place because the code did
not consistently use "reason" when appropriate).

BUG=None
TEST=All unit tests are passing.

Change-Id: Ic4fa3d80c06a2a8607e4961455174e067ccb6bed
Reviewed-on: https://chromium-review.googlesource.com/1704342
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>