cryptohome: remove SHA1 check in DecryptScrypt
The SHA1 MAC here is redundant since scryptenc_buf already
appends a MAC to the result buffer. Thus, the only way we could
fail checking this extra hash without failing first in
DecryptScryptBlob is if the hash itself was somehow tampered
with or corrupted, which doesn't have any bearing on the
integrity of the decrypted blob, so it can be safely ignored.

BUG=chromium:984782
TEST=unit tests, deploy, attempt to mount vault using scrypt
  successfully and also with tampered vault keyset

Change-Id: I5318a2454e43c6f2c103b3c2545a2589cd71efa0
Reviewed-on: https://chromium-review.googlesource.com/1703936
Tested-by: Greg Kerr <kerrnel@chromium.org>
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>