biod: Add support for migration of templates to have positive match secret.
If template needs migration, take the validation value that comes at the
next match. For template format compatibility, insert empty positive match
salt when uploading V3 templates to FPMCU firmware with template format V4.

BUG=chromium:927095
TEST=FEATURES=test emerge-nocturne biod
TEST=tested enrollment, matching, deleting and multifinger on nocturne
DUT with and without positive match secret support.
TEST=enrolled templates without positive match secret support, updated biod
and FPMCU to support positive match secret, verified that templates are
migrated to have positive match secret once there is a match, and continue to
work in further matches.
TEST=enrolled templates without positive match secret support, updated biod
(but not FPMCU) to support positive match secret, verified that templates
are not migrated and still work in matching.
TEST=enrolled template without positive match secret support, updated biod
and FPMCU to support positive match secret, but make FPMCU return an error
when asked for a template. Verified that biod obtained validation values for
the v3 templates on match, but the validation values were not written to
record (because biod was unable to get positive match salt with the template
and migration was not complete), and when biod restarts, these templates were
still v3 and marked "need migration".

Change-Id: I73aaa6e9c7200fee80d848e2d089d344bd4ec22e
Signed-off-by: Yicheng Li <yichengli@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1711107
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>