arc: networkd: Begin refactoring to remove ip helper process.
This patch is the first in a series to migrate all datapath
configuration into the main process. It introduces the guest
service construct and adds a placeh-holder implementation
for the existing ARC++ code, and wires in the lifecycle events,

*** This is a reland of https://crrev.com/c/1707789 but with the
guest event domain socket added back in which was removed too early and
broke adb proxying for arcvm.

BUG=b:133193765
BUG=b:133194026
BUG=b:136695230
TEST=units

Change-Id: I3d6db24e0f3158197f8c50fe42bfa647aa76c5f6
Reviewed-on: https://chromium-review.googlesource.com/1828657
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Ready: Garrick Evans <garrick@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>