libbrillo: Extract ScopedMountNamespace from arc-setup.
We'll be using a similar class in other places in platform2 so extract
this to avoid duplication.

Also add a couple of HANDLE_EINTR wrappers and shorten the name of the
factory method for creating ScopedMountNamespace objects.

BUG=chromium:985492
TEST=Compiles, no change in behaviour.

Change-Id: I4d9152bbd3c6f9cd7ad253f17fdc0ab38bbf2d9c
Reviewed-on: https://chromium-review.googlesource.com/1709488
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>