shill: Make Technology a fully-fledged class
Previously, Technology was not a real class, but instead essentially
served just as a namespace for static functions and an enum. Any
clients that wanted to use a "Technology" had to use a
Technology::Identifier instead. This change makes a Technology
instance a useful object that can be passed around.

BUG=None
TEST=All unit tests are passing.

Change-Id: I52810c9ad278f5190070eb1c257547f62034375c
Reviewed-on: https://chromium-review.googlesource.com/1708003
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>