crosh: Add gesture_prop command
This command is for interfacing with the gesture properties service. See
the design doc (go/cros-gesture-properties-dbus-design, internal only,
sorry) for more details on the service as a whole.

Due to the way parameters are split in crosh, it seems that we have to
have separate logic in the command itself for detecting and respecting
quotes around the property name.

BUG=chromium:952054
TEST=On a Samus, list devices, list the properties of a mouse, then get
     and set its "Mouse CPI" property (observing the resulting change in
     pointer speed. (Also test with --dash switch to crosh.)

Change-Id: I75258d3d9b02b0b5f5cd5604231af63945e861af
Reviewed-on: https://chromium-review.googlesource.com/1721522
Tested-by: Harry Cutts <hcutts@chromium.org>
Commit-Ready: Harry Cutts <hcutts@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>