camera: Split ZSL capture requests
This CL splits one ZSL capture request into two.

When VCD or Android camera framework sends a still capture request, a
preview output buffer usually comes with it. In the case of ZSL, we need
to put a raw input buffer in to reprocess it into the still output.
We handle this in adapter by splitting the request into 2 - a request
with the preview output buffer and a request with the still output
buffer. The latter one (still capture) would be transformed following
ZSL's mechanism, while the former one (preview) would have a ZSL ring
buffer attached to it.

BUG=b:130851309
TEST=Tested on kukui and still capture with ZSL works.

Change-Id: I7189c422d9bc9cc74ce9583a10d83ac2d3cec95a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1877636
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Commit-Queue: Jasmine Chen <lnishan@google.com>
Tested-by: Jasmine Chen <lnishan@google.com>