cryptohome: Add PopulateError helper.
This adds a helper function to populate the error pointer if it is not
null. This is necessary because cryptohome is full of nested if
statements to check the error pointer, which needlessly bloat the code
and reduce readability.

BUG=chromium:969193
TEST=emerge cryptohome

Change-Id: Ib6d181eea2f602dd260692fcc0c7bf5e2398b375
Reviewed-on: https://chromium-review.googlesource.com/1715101
Tested-by: Greg Kerr <kerrnel@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>