vm_tools: Make X sommelier scale internal displays differently
X11 applications historically use DPI to decide their scaling, however
this does not work for laptops, which are closer to you and so usually
have a higher DPI. To solve this problem, the X server usually lies
about the DPI of laptops.

We emulate this behaviour for chromeos by applying a scale factor if the
display is the internal one. We choose 0.66 semi-arbitrarily (based on
measuring sample laptop/desktop screen distances in our office).

Whilst this fixes some devices, it breaks others, base on a pre-existing
bug in sommelier where it would choose a DPI bucket that is no higher
than the true DPI. E.g. the pixelbook2's DPI of 232 would be rounded to
160 (not 240) which looks correct, because we werent applying the
laptop-distance scale, whereas another device's 168 DPI would be rounded
to 160, which looks unusually big (96 would be better), because we
werent applying the laptop-distance factor which means the "ideal" DPI
for that device is actually 112, closer to 96.

BUG=chromium:965817, chromium:973256, chromium:969462
TEST=/usr/share/code/code; \
     /opt/android-studio/bin/studio.sh

Change-Id: I55d2b09cf72bfbcde8dce135dd239de843cd802a
Reviewed-on: https://chromium-review.googlesource.com/1716347
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Ready: Nic Hollingum <hollingum@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>