tpm_manager: adopt Overalls for context connection part
This is in order to enable unittest to the modified part. Since the rest
of the logic could be re-work significantly in the near future, keeping
as is for now.
Note: this is a transient CL so it is expected that there is no test
added altogether.

BUG=b:138288163
BUG=b:137707011
TEST=build OK
TEST=tpm_manager_client read_space --index=0x10000001 --file=/dev/null

Cq-Depend: chromium:1717070
Change-Id: I5ed428b0d396c613b20b1fb980e3cb7d5a8054e4
Reviewed-on: https://chromium-review.googlesource.com/1716773
Tested-by: Leo Lai <cylai@google.com>
Commit-Ready: Leo Lai <cylai@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>