libbrillo: remove SetReceiveBufferSize from UdevMonitor
The default size of the socket receive buffer is a few hundred
KB and I don't think any users of this library are going to want
to change it, but it is easy to misuse by setting to a value
which is too low. Let's remove it for now and we can always add
it back in later if we need to.

BUG=chromium:986484
TEST=libbrillo and mist unit tests

Change-Id: I371b9ac05062285215bc7aa3adc69aae9baf3440
Reviewed-on: https://chromium-review.googlesource.com/1716776
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Anand Mistry <amistry@chromium.org>