libmems: Reset Buffer When Error Occurs
This commit fixes the bug that when error occurs, the deprecated buffer
should be reset.

A potential error: when the refill function gets a timeout or other
error, the file descriptor that the buffer uses will be invalidated.
Therefore, we should reset the buffer to prevent it being reused in the
next ReadEvents.

BUG=chromium:971175
TEST=build with debug code in mems_setup/main.cc and camera/hal_adapter
     observe events are retrieved

Change-Id: I28a5d85fd866aa35e12a4756d7886f3219a11e26
Reviewed-on: https://chromium-review.googlesource.com/1718284
Tested-by: Cheng-Hao Yang <chenghaoyang@chromium.org>
Tested-by: Enrico Granata <egranata@chromium.org>
Commit-Ready: Cheng-Hao Yang <chenghaoyang@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Enrico Granata <egranata@chromium.org>