diagnostics: Change structure of EcEvent
Before, the payload of an EcEvent was just an opaque array.
To allow for parsing the event for certain flags which should
cause notifications, re-structure the payload array to
packed structs.

Also fix the meaning of the |size| field in EcEvent.
Before, the |size| field of an EcEvent was interpreted to mean
"The number of words in the payload array." This is incorrect,
the EC will fill this field with "The number of following words
in the struct", AKA including the |type| word, AKA one more than
we thought. With the incorrect interpretation, some valid EcEvents
are filtered out from being sent in
WilcoDtcSupportdCore::SendGrpcEcEventToWilcoDtc. This patch
corrects the interpretation of |size| when we construct EcEvents
and decode them for making gRPC calls.

TEST=unit tests
BUG=b:126320709, chromium:972105

Change-Id: Idd69f89682be0b10723465ef34a4687c6ef0508c
Reviewed-on: https://chromium-review.googlesource.com/1705336
Tested-by: Nick Crews <ncrews@chromium.org>
Commit-Ready: Nick Crews <ncrews@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Trent Begin <tbegin@chromium.org>
Reviewed-by: Nick Crews <ncrews@chromium.org>