power: move SensorRole to UserProximityObserver
Whether a sensor is uses for detecting WiFi or LTE proximity
is not a feature of SAR, but a generic configuration parameter
for a proximity sensor.

This is part of an initial round of refactoring aimed at supporting
sources of proximity information other than a SAR sensor in powerd.

BUG=b:123434029
TEST=ran powerd unit tests

Change-Id: I36d9275aa2317edd8e91c59f3c77eec152e14bd8
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1720863
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>