cryptohome: Extract mount operations into a helper class.
The road to a better cryptohome starts with a single CL.

This CL extracts all functions that result in calls to the mount(2)
syscall into a helper class. The way the existing code in mount.cc was
structured made the decision of where to "cut" pretty easy: we have
public Mount{Ephemeral}Cryptohome functions, and private
Mount{Ephemeral}CryptohomeInner helper functions. Extract
MountEphemeralCryptohomeInner and its dependencies into the helper
class.

MountCryptohomeInner is less trivial to partition, so extracting it
fully into the helper class (or exposing a single function in the
helper class that carries out all the mounts required for non-ephemeral
cryptohomes) will be done in the CL immediately following this one.
Besides, this CL is already large as-is.

To avoid adding functional changes to a big refactoring CL, this CL
does not change where state is stored, and keeps the new helper class
fully stateless. This results in some objects being passed into the
helper as pointers, but with the helper class having no state and the
helper class instances not living beyond a single mount.cc function,
this is not problematic. The third CL in this effort will resolve the
situation of where state is kept.

An overview of how the work has been staged can be seen at
https://docs.google.com/document/d/1hSUSuIQLZvH1OVwn7N8lDIT7lFUEeHkmFDXQduBccho/edit.
This document also includes justifications for splitting the CLs this
way, and for extracting this functionality and not more (or less).

The main design document for this work is go/cros-non-root-mount-ns.

While this CL appears large, there are no functional changes. It's just
code being moved around, plus necessary changes for unit tests to still
compile and pass.

BUG=chromium:985492
TEST=Unit tests pass.
TEST=Entering and exiting guest mode works, no cryptohome error messages.
TEST=Adding a new user works, no cryptohome error messages.
TEST=Logging in to an existing user account works. Data is preserved.

Change-Id: I2679c88470da598767684c4c10aabff5a993adeb
Reviewed-on: https://chromium-review.googlesource.com/1729438
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>