shill: connection: Update default route metrics along with policy rules
Connection::SetMetric is called by Manager to inform a Connection
instance of a new metric. The primary effect of this call is to modify
the priority of policy routing rules. However, since the metric of a
Connection also has an effect on the metric of newly-created default
routes, we would like to be consistent and also update the metric of
existing default routes when SetMetric is called.

BUG=chromium:941597
TEST=-All unit tests are passing.
     -Manual testing, checking route metrics with
      `ip route show table 0 | grep '^default .* metric'`.

Change-Id: I1c5e62432c1ac7bfc67e48eca09a948538cf0435
Reviewed-on: https://chromium-review.googlesource.com/1730177
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>