shill: routing_table_entry: Remove from_rtnl field
Previously, the from_rtnl field was set when a route was learned about
by listening to rtnl messages (rather than coming from a client within
shill). This distinction provides little value, and the code currently
performs no useful logic with regards to whether or not the field is
set. The primary user of from_rtnl is test code, which explicitly sets
from_rtnl to fake that a RoutingTableEntry came from
RoutingTable::ParseRoutingTableMessage.

If in the future we would like to distinguish between routes from
different clients or for different purposes, the protocol field would
be a more useful, appropriate, and kernel/rtnl-supported means of
doing so.

BUG=None
TEST=All unit tests are passing.

Change-Id: Ic21f3643f62d5f0613f939ca4b828782aa40274a
Reviewed-on: https://chromium-review.googlesource.com/1731350
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>