diagnostics: refine probe mojo interface
This CL proposes a new API that cros_healthd would expose
via Mojo.

Now that cros_healthd is close to being able to replace
runtime_probe in device_status_collector, we want to propose
an API that will satisfy developers working on cros_healthd,
security and privacy teams reviewing new additions to the
data cros_healthd exposes, and developers working in Chromium
and the cloud who will use the data exposed by cros_healthd.

From a developer perspective, we want an API that is easy to
extend. Here, if we wanted to add a new category of data, say
graphics-related information, we would just need to add a new
ProbeCategoryEnum, a new GraphicsInfo struct, and then add a
nullable GraphicsInfo to TelemetryInfo.

In terms of development in the stack above cros_healthd, no
new changes would be required when a new category of data was
added to cros_healthd, unless the application needs to access
that new data.

From a security standpoint, all of the data is fully typed,
and there is no multiplexing or parsing involved.

BUG=b:128683357
TEST=none

Cq-Depend:chromium:1779336
Change-Id: I7b4492c83cf60627e010780c6eac7c2ab14358f4
Signed-off-by: Paul Moy <pmoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1732800
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Kartik Hegde <khegde@chromium.org>