debugd: Remove anonymizer from debugd
The path used for this anonymizer duplicates work with the anonymizer in
Chrome and the anonymizer in Chrome has been updated so that it covers
all cases the one in debugd was doing as well. Now we can remove this
one so the anonymizer is in one place.

BUG=chromium:921388
TEST=Verified feedback report is still anonymized

Change-Id: If1e9dcb9015e4ecf2734cc5ff042e1ca23f1bf69
Reviewed-on: https://chromium-review.googlesource.com/1735515
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>