cryptohome: Expose mount operations in a single function.
In order to move cryptohome mount operations out of process, first we
need to expose them as a single function in the helper. The reason for
this is pretty straightforward: when executing out of process, the
helper needs to be able to carry out all the mount operations in a
single invocation. The easiest way to ensure that this will be possible
is to perform all mount operations as a result of a single function
invocation. Then we change the function invocation to a helper
executable invocation and we're done.

This continues the work in the previous CL by transferring enough state
to the MountHelper class to allow regular user mounts to be completed
with a single call to the helper. In order to make the helper API less
overly verbose, this CL also introduces an Options struct for the
helper.

As opposed to ephemeral mounts, where the entire *Inner function could
be moved to the helper, for regular mounts the *Inner function does
quite a bit of key manipulation. Per the design these should stay in
cryptohome and not be done in the helper, so for regular mounts the
helper calls happen inside of the *Inner function instead of fully
replacing it. This is not a big deal since seam between the Mount class
and the helper is still pretty obvious since key operations happen
first and mount operations happen later.

The next CL in the series will move unmount calls to the helper, in
preparation for running things out of process (this is the plan at
https://docs.google.com/document/d/1hSUSuIQLZvH1OVwn7N8lDIT7lFUEeHkmFDXQduBccho/edit#.)

This CL is a net addition of fewer than 100 lines. Most of it is code
moved around.

BUG=chromium:985492
TEST=Unit tests pass.
TEST=Add new user to Chromebook successfully.
TEST=Log in, create local data, log out.
TEST=Log back in, local data is still present.

Change-Id: I58607fe2f9962f973b94b793156849e085f327df
Reviewed-on: https://chromium-review.googlesource.com/1769114
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>