vm_tools: Remove Concierge process itself from /sys/fs/cgroup/cpu/vms
Previously, the process was added to the cpu cgroup just to have all
the crosvm processes in the group. With CL:1742996, we can do the
same without adding Concierge itself to the group.

This CL also makes it easier to add each VM (Termina, Plugin, ...)
to a different cgroup in the future (b/139095697).

BUG=b:139095697
TEST=Start Concierge, verify /sys/fs/cgroup/cpu/vms/cpu.share is 896
TEST=Start Concierge, verify /sys/fs/cgroup/cpu/vms/tasks is empty
TEST=Start Crostini, verify /sys/fs/cgroup/cpu/vms/tasks is not empty

Change-Id: I4e4d04e68e28afbc8b815d4f1bbdb687d92c7325
Reviewed-on: https://chromium-review.googlesource.com/1745828
Tested-by: Yusuke Sato <yusukes@chromium.org>
Commit-Ready: Yusuke Sato <yusukes@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>