shill: connection: Remove Binder
Connection::Binder was previously used as a way of informing
dependents when a Connection goes down. However, there is currently no
code in Shill that actually needs this. The only Binder client is
VPNService, but this usage is unnecessary. The lifetime of a
Connection is controlled by Device, which will call
SetConnection(nullptr) when removing an existing Connection, so
VPNService can cleanly keep track of Connection removals without the
use of Connection::Binder.

Functionality will be added to VirtualDevice to map virtual devices
onto physical devices. Doing this in Device rather than Connection
allows for VirtualDevices to map to a physical Device even when the
Device is not yet up (e.g. arc_wlan0 should still map to wlan0 even
when we have not yet connected to a WiFiService). This also allows us
to expose Device dependency information to the D-Bus
interface (e.g. if we have both a cellular dongle and l2tp VPN, there
is currently no clean way to determine which PPPDevice maps to which
usage from the D-Bus interface).

BUG=b:138464797
TEST=All unit tests are passing.

Change-Id: Ic23eaa66646b9bdab614be7bca04a77179bfb477
Reviewed-on: https://chromium-review.googlesource.com/1748120
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>