dlcservice: Error and Status propagation
Error and status are propagated up according to the constants set in
system_api. The examples and dlcservice_util reflect these changes and
how to utilize the changed flow of working with dlcservice.

Examples and dlcservice_util reflect use of changes.

Reording of fields within .proto is safe as this is not used yet in
production with no requirement to be backwards compatible.

BUG=chromium:993559
TEST=FEATURES="test" emerge-$BOARD dlcservice dlcservice-client system_api
TEST=tast run $TEST_IP platform.DLCService

Change-Id: I27f56ee821ed0f01dacb24d707874fc6b9a2aba0
Reviewed-on: https://chromium-review.googlesource.com/1756583
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Ready: Jae Hoon Kim <kimjae@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>