power: check wakeup count again when canceling suspend
This allows powerd_suspend to signal a suspend as failed when
the kernel returns EBUSY but we didn't see an increased wake
count. This could happen, for example, if userspace processes
fail to freeze, and it is a different situation than a late
cancellation via wake event so we should log it differently.

BUG=b:139444006
TEST=inject EBUSY fault into suspend path and verify that it
  is marked as failed unless the wakeup count increased

Change-Id: I5a694db4b49b227b082561dd2dfe070ca0d9db4e
Reviewed-on: https://chromium-review.googlesource.com/1754863
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>