u2fd: Remove version command
This command was only used to simulate latency
in the implementation of the PING HID command.

The version command is being removed from the
firmware, so we replace it with another command
that is still supported.

BUG=b:133275654
TEST=build, deploy, check U2F can be used to
     register a new credential and log in

Change-Id: Iddacdc0c18427faac566e29760749fa4c98614b4
Reviewed-on: https://chromium-review.googlesource.com/1755862
Tested-by: Louis Collard <louiscollard@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>