bluetooth: Add a getter function of connection ID based on device address
In libnewblue, GATT APIs take a connection ID as a parameter to perform GATT
operation and pass the result along with the connection ID. And
gatt_interface_handle passes device address as a parameter to gatt.
Therefore, in order for gatt to retrieve connection ID based on device
address, device_interface_handler should provide a getter for it.

BUG=b:134425062
TEST=None

Change-Id: I5c9c14d0a56e196c3597a156ef217c69c5c47ea0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1756933
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Yu Liu <yudiliu@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>